add : 데이터, ui 컴포넌트 분리
