+++ src/refactoring/App.tsx
@@ -1,59 +1,15 @@
 import { useState } from 'react';
 import { AdminPage } from './page/admin';
 import { CartPage } from './page/cart';
-
-import { Coupon, Product } from './features/shared/types/entities.ts';
 import { useCoupons, useProducts } from './features/cart/hooks';
-
-// 초기 상품 목록
-const initialProducts: Product[] = [
-  {
-    id: 'p1',
-    name: '상품1',
-    price: 10000,
-    stock: 20,
-    discounts: [
-      { quantity: 10, rate: 0.1 },
-      { quantity: 20, rate: 0.2 },
-    ],
-  },
-  {
-    id: 'p2',
-    name: '상품2',
-    price: 20000,
-    stock: 20,
-    discounts: [{ quantity: 10, rate: 0.15 }],
-  },
-  {
-    id: 'p3',
-    name: '상품3',
-    price: 30000,
-    stock: 20,
-    discounts: [{ quantity: 10, rate: 0.2 }],
-  },
-];
-
-// 초기 쿠폰 목록
-const initialCoupons: Coupon[] = [
-  {
-    name: '5000원 할인 쿠폰',
-    code: 'AMOUNT5000',
-    discountType: 'amount',
-    discountValue: 5000,
-  },
-  {
-    name: '10% 할인 쿠폰',
-    code: 'PERCENT10',
-    discountType: 'percentage',
-    discountValue: 10,
-  },
-];
+import { INITIAL_PRODUCTS } from './constants/initialProducts';
+import { INITIAL_COUPONS } from './constants/initialCoupons';
 
 const App = () => {
   // 상품 관리 훅 사용
-  const { products, updateProduct, addProduct } = useProducts(initialProducts);
+  const { products, updateProduct, addProduct } = useProducts(INITIAL_PRODUCTS);
   // 쿠폰 관리 훅 사용
-  const { coupons, newCoupon, setNewCoupon, handleAddCoupon } = useCoupons(initialCoupons);
+  const { coupons, newCoupon, setNewCoupon, handleAddCoupon } = useCoupons(INITIAL_COUPONS);
   // 관리자 페이지 표시 여부 상태 관리
   const [isAdmin, setIsAdmin] = useState(false);
