+++ src/refactoring/constants/index.ts
@@ -1,0 +1,3 @@
+export * from './initialProducts';
+export * from './initialCoupons';
+export * from './productFormFields';
