+++ src/refactoring/constants/initialCoupons.ts
@@ -1,0 +1,16 @@
+import { Coupon } from '../features/shared/types/entities';
+
+export const INITIAL_COUPONS: Coupon[] = [
+  {
+    name: '5000원 할인 쿠폰',
+    code: 'AMOUNT5000',
+    discountType: 'amount',
+    discountValue: 5000,
+  },
+  {
+    name: '10% 할인 쿠폰',
+    code: 'PERCENT10',
+    discountType: 'percentage',
+    discountValue: 10,
+  },
+];
