+++ src/refactoring/constants/initialProducts.ts
@@ -1,0 +1,28 @@
+import { Product } from '../features/shared/types/entities';
+
+export const INITIAL_PRODUCTS: Product[] = [
+  {
+    id: 'p1',
+    name: '상품1',
+    price: 10000,
+    stock: 20,
+    discounts: [
+      { quantity: 10, rate: 0.1 },
+      { quantity: 20, rate: 0.2 },
+    ],
+  },
+  {
+    id: 'p2',
+    name: '상품2',
+    price: 20000,
+    stock: 20,
+    discounts: [{ quantity: 10, rate: 0.15 }],
+  },
+  {
+    id: 'p3',
+    name: '상품3',
+    price: 30000,
+    stock: 20,
+    discounts: [{ quantity: 10, rate: 0.2 }],
+  },
+];
