+++ src/refactoring/features/cart/components/index.ts
@@ -1,2 +1,6 @@
 export * from './cart-detail';
 export * from './proudct-list';
+export * from './product-card';
+export * from './cart-item-card';
+export * from './coupon-selector';
+export * from './order-summary';
